Adiciona Domain/Entities/DetalhePedido.cs

+++ src/Domain/Entities/DetalhePedido.cs
@@ -4,8 +4,8 @@ namespace Ecommerce.Domain.Entities
     {
         public string? NomeProduto { get; set; }
         public int Quantidade { get; set; }
-        public double Preco { get; set; }
-        public double SubTotal { get; set; }
+        public decimal Preco { get; set; }
+        public decimal SubTotal { get; set; }
         public Pedido? Pedido { get; set; }
         public int PedidoId { get; set; }
     }
